Распилил код на несколько файлов, для большей удобочитаемости
