+++ samples/strategy/SimUDuck/SimUDuck.cpp
@@ -1,178 +1,10 @@
-#include <cassert>
-#include <iostream>
-#include <memory>
-#include <vector>
-
-using namespace std;
-
-struct IFlyBehavior
-{
-	virtual ~IFlyBehavior(){};
-	virtual void Fly() = 0;
-};
-class FlyWithWings : public IFlyBehavior
-{
-public:
-	void Fly() override
-	{
-		cout << "I'm flying with wings!!" << endl;
-	}
-};
-
-class FlyNoWay : public IFlyBehavior
-{
-public:
-	void Fly() override {}
-};
-
-struct IQuackBehavior
-{
-	virtual ~IQuackBehavior(){};
-	virtual void Quack() = 0;
-};
-
-class QuackBehavior : public IQuackBehavior
-{
-public:
-	void Quack() override
-	{
-		cout << "Quack Quack!!!" << endl;
-	}
-};
-class SqueakBehavior : public IQuackBehavior
-{
-public:
-	void Quack() override
-	{
-		cout << "Squeek!!!" << endl;
-	}
-};
-
-class MuteQuackBehavior : public IQuackBehavior
-{
-public:
-	void Quack() override {}
-};
-
-class Duck
-{
-public:
-	Duck(unique_ptr<IFlyBehavior>&& flyBehavior,
-		unique_ptr<IQuackBehavior>&& quackBehavior)
-		: m_quackBehavior(move(quackBehavior))
-	{
-		assert(m_quackBehavior);
-		SetFlyBehavior(move(flyBehavior));
-	}
-	void Quack() const
-	{
-		m_quackBehavior->Quack();
-	}
-	void Swim()
-	{
-		cout << "I'm swimming" << endl;
-	}
-	void Fly()
-	{
-		m_flyBehavior->Fly();
-	}
-	virtual void Dance()
-	{
-		cout << "I'm Dancing" << endl;
-	}
-	void SetFlyBehavior(unique_ptr<IFlyBehavior>&& flyBehavior)
-	{
-		assert(flyBehavior);
-		m_flyBehavior = move(flyBehavior);
-	}
-	virtual void Display() const = 0;
-	virtual ~Duck() = default;
-
-private:
-	unique_ptr<IFlyBehavior> m_flyBehavior;
-	unique_ptr<IQuackBehavior> m_quackBehavior;
-};
-
-class MallardDuck : public Duck
-{
-public:
-	MallardDuck()
-		: Duck(make_unique<FlyWithWings>(), make_unique<QuackBehavior>())
-	{
-	}
-
-	void Display() const override
-	{
-		cout << "I'm mallard duck" << endl;
-	}
-};
-
-class RedheadDuck : public Duck
-{
-public:
-	RedheadDuck()
-		: Duck(make_unique<FlyWithWings>(), make_unique<QuackBehavior>())
-	{
-	}
-	void Display() const override
-	{
-		cout << "I'm redhead duck" << endl;
-	}
-};
-class DecoyDuck : public Duck
-{
-public:
-	DecoyDuck()
-		: Duck(make_unique<FlyNoWay>(), make_unique<MuteQuackBehavior>())
-	{
-	}
-	void Display() const override
-	{
-		cout << "I'm decoy duck" << endl;
-	}
-	void Dance() override {}
-};
-class RubberDuck : public Duck
-{
-public:
-	RubberDuck()
-		: Duck(make_unique<FlyNoWay>(), make_unique<SqueakBehavior>())
-	{
-	}
-	void Display() const override
-	{
-		cout << "I'm rubber duck" << endl;
-	}
-	void Dance() override {}
-};
-
-class ModelDuck : public Duck
-{
-public:
-	ModelDuck()
-		: Duck(make_unique<FlyNoWay>(), make_unique<QuackBehavior>())
-	{
-	}
-	void Display() const override
-	{
-		cout << "I'm model duck" << endl;
-	}
-	void Dance() override {}
-};
-
-void DrawDuck(Duck const& duck)
-{
-	duck.Display();
-}
-
-void PlayWithDuck(Duck& duck)
-{
-	DrawDuck(duck);
-	duck.Quack();
-	duck.Fly();
-	duck.Dance();
-	cout << endl;
-}
+#include "lib/Duck/DecoyDuck.h"
+#include "lib/Duck/MallardDuck.h"
+#include "lib/Duck/ModelDuck.h"
+#include "lib/Duck/RedheadDuck.h"
+#include "lib/Duck/RubberDuck.h"
+#include "lib/DuckFunctions.h"
+#include <cstdlib>
 
 int main()
 {
@@ -190,6 +22,9 @@ int main()
 
 	ModelDuck modelDuck;
 	PlayWithDuck(modelDuck);
-	modelDuck.SetFlyBehavior(make_unique<FlyWithWings>());
+
+	modelDuck.SetFlyBehavior(std::make_unique<FlyWithWings>());
 	PlayWithDuck(modelDuck);
+
+	return EXIT_SUCCESS;
 }
