+++ samples/strategy/SimUDuck/lib/Duck/Duck.h
@@ -1,0 +1,57 @@
+#ifndef DUCK_H
+#define DUCK_H
+
+#include "Fly/IFlyBehavior.h"
+#include "Quack/IQuakBehavior.h"
+
+#include <cassert>
+#include <iostream>
+#include <memory>
+#include <vector>
+
+class Duck
+{
+public:
+	Duck(std::unique_ptr<IFlyBehavior>&& flyBehavior,
+		std::unique_ptr<IQuackBehavior>&& quackBehavior)
+		: m_quackBehavior(std::move(quackBehavior))
+	{
+		assert(m_quackBehavior);
+		SetFlyBehavior(std::move(flyBehavior));
+	}
+
+	void Quack() const
+	{
+		m_quackBehavior->Quack();
+	}
+
+	void Swim()
+	{
+		std::cout << "I'm swimming" << std::endl;
+	}
+
+	void Fly()
+	{
+		m_flyBehavior->Fly();
+	}
+
+	virtual void Dance()
+	{
+		std::cout << "I'm Dancing" << std::endl;
+	}
+
+	void SetFlyBehavior(std::unique_ptr<IFlyBehavior>&& flyBehavior)
+	{
+		assert(flyBehavior);
+		m_flyBehavior = std::move(flyBehavior);
+	}
+
+	virtual void Display() const = 0;
+	virtual ~Duck() = default;
+
+private:
+	std::unique_ptr<IFlyBehavior> m_flyBehavior;
+	std::unique_ptr<IQuackBehavior> m_quackBehavior;
+};
+
+#endif
