+++ samples/strategy/SimUDuck/lib/Duck/Fly/FlyNoWay.h
@@ -1,0 +1,12 @@
+#ifndef FLYNOWAY_H
+#define FLYNOWAY_H
+
+#include "IFlyBehavior.h"
+
+class FlyNoWay : public IFlyBehavior
+{
+public:
+	void Fly() override {}
+};
+
+#endif
