+++ samples/strategy/SimUDuck/lib/Duck/Fly/FlyWithWings.h
@@ -1,0 +1,16 @@
+#ifndef FLYWITHWINGS_H
+#define FLYWITHWINGS_H
+
+#include "IFlyBehavior.h"
+#include <iostream>
+
+class FlyWithWings : public IFlyBehavior
+{
+public:
+	void Fly() override
+	{
+		std::cout << "I'm flying with wings!!" << std::endl;
+	}
+};
+
+#endif
