+++ samples/strategy/SimUDuck/lib/Duck/Fly/IFlyBehavior.h
@@ -1,0 +1,10 @@
+#ifndef LAB01_SIMUDUCK_IFLYBEHAVIOR_H
+#define LAB01_SIMUDUCK_IFLYBEHAVIOR_H
+
+struct IFlyBehavior
+{
+	virtual ~IFlyBehavior(){};
+	virtual void Fly() = 0;
+};
+
+#endif
