+++ samples/strategy/SimUDuck/lib/Duck/MallardDuck.h
@@ -1,0 +1,22 @@
+#ifndef MALLARDDUCK_H
+#define MALLARDDUCK_H
+
+#include "Duck.h"
+#include "Fly/FlyWithWings.h"
+#include "Quack/QuackBehavior.h"
+
+class MallardDuck : public Duck
+{
+public:
+	MallardDuck()
+		: Duck(std::make_unique<FlyWithWings>(), std::make_unique<QuackBehavior>())
+	{
+	}
+
+	void Display() const override
+	{
+		std::cout << "I'm mallard duck" << std::endl;
+	}
+};
+
+#endif
