+++ samples/strategy/SimUDuck/lib/Duck/Quack/IQuakBehavior.h
@@ -1,0 +1,10 @@
+#ifndef IQUAKBEHAVIOR_H
+#define IQUAKBEHAVIOR_H
+
+struct IQuackBehavior
+{
+	virtual ~IQuackBehavior(){};
+	virtual void Quack() = 0;
+};
+
+#endif
