+++ samples/strategy/SimUDuck/lib/Duck/Quack/MuteQuackBehavior.h
@@ -1,0 +1,12 @@
+#ifndef MUTEQUAKBEHAVIOR_H
+#define MUTEQUAKBEHAVIOR_H
+
+#include "IQuakBehavior.h"
+
+class MuteQuackBehavior : public IQuackBehavior
+{
+public:
+	void Quack() override {}
+};
+
+#endif
