+++ samples/strategy/SimUDuck/lib/Duck/Quack/SqueakBehavior.h
@@ -1,0 +1,16 @@
+#ifndef SQUEAKBEHAVIOR_H
+#define SQUEAKBEHAVIOR_H
+
+#include "IQuakBehavior.h"
+#include <iostream>
+
+class SqueakBehavior : public IQuackBehavior
+{
+public:
+	void Quack() override
+	{
+		std::cout << "Squeek!!!" << std::endl;
+	}
+};
+
+#endif
