+++ samples/strategy/SimUDuck/lib/Duck/RedheadDuck.h
@@ -1,0 +1,24 @@
+#ifndef REDHEADDUCK_H
+#define REDHEADDUCK_H
+
+#include "Duck.h"
+#include "Fly/FlyWithWings.h"
+#include "Quack/QuackBehavior.h"
+
+#include <memory>
+
+class RedheadDuck : public Duck
+{
+public:
+	RedheadDuck()
+		: Duck(std::make_unique<FlyWithWings>(), std::make_unique<QuackBehavior>())
+	{
+	}
+
+	void Display() const override
+	{
+		std::cout << "I'm redhead duck" << std::endl;
+	}
+};
+
+#endif
