+++ samples/strategy/SimUDuck/lib/Duck/RubberDuck.h
@@ -1,0 +1,25 @@
+#ifndef RUBBERDUCK_H
+#define RUBBERDUCK_H
+
+#include "Duck.h"
+#include "Fly/FlyNoWay.h"
+#include "Quack/SqueakBehavior.h"
+#include <iostream>
+
+class RubberDuck : public Duck
+{
+public:
+	RubberDuck()
+		: Duck(std::make_unique<FlyNoWay>(), std::make_unique<SqueakBehavior>())
+	{
+	}
+
+	void Display() const override
+	{
+		std::cout << "I'm rubber duck" << std::endl;
+	}
+
+	void Dance() override {}
+};
+
+#endif
