+++ samples/strategy/SimUDuck/lib/DuckFunctions.h
@@ -1,0 +1,20 @@
+#ifndef DUCKFUNCTIONS_H
+#define DUCKFUNCTIONS_H
+
+#include "Duck/Duck.h"
+
+void DrawDuck(Duck const& duck)
+{
+	duck.Display();
+}
+
+void PlayWithDuck(Duck& duck)
+{
+	DrawDuck(duck);
+	duck.Quack();
+	duck.Fly();
+	duck.Dance();
+	std::cout << std::endl;
+}
+
+#endif
